+++ Sources/CommandRegistry/TypeAliases.swift
@@ -1,10 +1,3 @@
-//
-//  TypeAliases.swift
-//  CommandRegistry
-//
-//  Created by Eneko Alonso on 3/9/19.
-//
-
 import Foundation
 import Utility
